Fixed a bug causing the shapes to not be able to move completly to the right.
Before in the allowed function I checked whether curx + block.x is greater or equal to NUMCOLS-1 (=9), which caused the function to already set the move invalid when it went to the last index of the row (9).
Consequently the bug occured. The fix is, to remove the -1 to get NUMCOLS.

The check for y is not affected as y is incremented by one, during the acces for shape setting to avoid a buffer underflow, due to -1 coordinates in the shape descriptions.