Done some little refactoring about the addressing in the board, to simplify the program.
The shapes are given in a coordinate system, with both axis starting from -1, which is great, because one can use the used rotation algorithms, but it's bad, because the negative index could lead to a buffer
underflow. So every y index is incremented by one when accesing the board array. Until now this has been done in every acces to the board, why soon led to some cunfusion, which is why now this has been resolved
by just putting a 1 in cury intially, rather than a 0.